requirements.txt
Django~=5.1.4
django-tailwind
mysqlclient
pillow
django-browser-reload
requests